requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
AnyQt==0.2.0
asttokens==2.2.1
attrs==23.2.0
backcall==0.2.0
backports.tarfile==1.1.1
baycomp==1.0.3
bio==1.5.3
biopython==1.80
biothings-client==0.2.6
blinker==1.8.2
Bottleneck==1.3.8
cachetools==5.3.3
catboost==1.2.5
cattrs==23.2.3
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dictdiffer==0.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==1.2.0
fastapi==0.110.2
ffmpy==0.3.2
filelock==3.13.4
fonttools==4.38.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
gradio==4.28.3
gradio_client==0.16.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==2.10
importlib-metadata==6.1.0
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.22.0
ipython==8.12.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.18.2
Jinja2==3.1.3
joblib==1.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.1.0
jupyter_core==5.3.0
keyring==25.2.0
keyrings.alt==5.0.1
kiwisolver==1.4.4
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==8.7.0
mygene==3.2.2
nest-asyncio==1.5.6
numba==0.58.1
numpy==1.26.4
openpyxl==3.0.10
openTSNE==1.0.1
orange-canvas-core==0.1.35
orange-widget-base==4.22.0
Orange3==3.36.2
Orange3-Associate==1.3.0
orjson==3.10.1
packaging==23.0
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
plotly==5.21.0
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==16.0.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
pydub==0.25.1
Pygments==2.14.0
pyparsing==3.0.9
pyqtgraph==0.13.6
python-dateutil==2.8.2
python-louvain==0.16
python-multipart==0.0.9
pytz==2021.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.0.2
qasync==0.27.1
qtconsole==5.5.1
QtPy==2.4.1
Quandl==3.6.1
referencing==0.35.0
requests==2.31.0
requests-cache==1.2.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.4.2
scikit-learn==1.3.2
scipy==1.13.0
semantic-version==2.10.0
serverfiles==0.3.1
shap==0.43.0
shellingham==1.5.4
six==1.15.0
sklearn==0.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.1
soundfile==0.11.0
stack-data==0.6.2
starlette==0.37.2
streamlit==1.34.0
tenacity==8.2.3
threadpoolctl==2.1.0
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
wcwidth==0.2.6
websockets==11.0.3
xgboost==2.0.1
xlrd==2.0.1
XlsxWriter==3.0.5
zipp==3.15.0
